+++ src/madeup/ExpressionBlock.cpp
@@ -30,7 +30,7 @@ Co<Expression> ExpressionBlock::evaluate(Environment &env) const {
 
 /* ------------------------------------------------------------------------- */
 
-Co<Expression> ExpressionBlock::operator[](int i) {
+Co<Expression> ExpressionBlock::operator[](size_t i) {
   /* throw 5; */
   /* std::cout << "i: " << i << std::endl; */
   assert(0 <= i && i < statements.size());
